fix(governance): harden action item completion receipts (review)
Addresses all 5 unresolved review threads on PR #1661:

1. Append-only completion-receipt history (Codex P1)
   ----------------------------------------------------
   The sled key scheme overwrote prior receipts on a reopen / re-complete
   cycle. Replaced with a primary-by-record_hash + secondary-by-item
   index pair:

       receipt:action_item_completion:rec:{record_hash} -> JSON
       receipt:action_item_completion:by_item:
           {u64 BE item_id_len}{item_id bytes}
           {u64 BE completed_at}{32-byte record_hash} -> ""

   `get_action_item_completion_by_item` returns the latest by
   completed_at via the by_item index. New trait method
   `list_action_item_completions_by_item` returns the full chain
   oldest-first; default impl is `Ok(vec![])`.

2. Fail completion when receipt persistence fails (Codex P1)
   ----------------------------------------------------------
   `update_action_item_status` now persists the receipt BEFORE saving
   the status. If the backend rejects the receipt the status save does
   not run — the holder's standing never advertises a completion
   that has no provenance. Receipts are no longer best-effort.

3. Full-update handler bypass (Copilot)
   --------------------------------------
   `PUT /v1/gov/domains/{domain_id}/action-items/{item_id}` could set
   `status=Completed` via `manager.update_action_item`, bypassing the
   receipt-emitting `update_action_item_status` path. The full-update
   handler now applies non-status fields via the original write path
   and routes any status change through `update_action_item_status`
   with the authenticated DID as actor.

4. Canonical hash length prefixes (Copilot)
   ------------------------------------------
   `ActionItemCompletionReceipt::compute_record_hash` used `u32` length
   prefixes; the rest of the module uses `u64`. Switched to `u64` so
   the receipt's canonical encoding is consistent with
   `GovernanceProof::compute_proof_hash` and `compute_vote_hash`.

5. Per-write sled flush (Copilot)
   --------------------------------
   `ReceiptStore::put_action_item_completion` called `db.flush()` per
   insert, unlike `put_governance` and other receipt writes. Removed
   so durability semantics match the rest of the receipt store.

## New tests

  - full_update_to_completed_emits_completion_receipt — exercises
    the HTTP handler with a `status: completed` body and asserts a
    receipt is persisted.
  - receipt_backend_failure_prevents_completed_status_commit — wires
    a `FailingCompletionStore` backend and asserts the action item's
    stored status does NOT advance to Completed when the receipt
    backend rejects.
  - reopen_and_recomplete_preserves_completion_history — sleeps 1.1s
    between completions so wall-clock `completed_at` advances; asserts
    the chain has 2 distinct receipts ordered oldest-first; asserts
    `get_action_item_completion_by_item` returns the latest.

## Validation

  - cargo fmt --all --check  clean
  - cargo clippy -p icn-governance-actor -p icn-governance -p icn-gateway
    --all-targets -- -D warnings  clean
  - cargo test -p icn-governance-actor --test me_action_item_receipt_chain
    6/6 pass (was 3, now 6)
  - cargo test -p icn-governance-actor --test me_action_card_receipt_chain
    3/3 pass (no regression)
  - cargo test -p icn-governance-actor --test me_action_cards
    4/4 pass (no regression)
  - cargo test -p icn-governance-actor --test me_standing
    9/9 pass (no regression)
  - cargo test -p icn-governance-actor --lib  220/220 pass
  - cargo test -p icn-governance --lib  502/502 pass
  - cargo test -p icn-gateway --lib  528/528 pass
  - python3 docs/scripts/doc_control_check.py --strict  pass
    (36 pre-existing warnings, none new)
  - git diff --check  clean
  - public-website forbidden-term grep  clean

Refs icn#1646.

+++ icn/crates/icn-governance/src/proof.rs
@@ -573,8 +573,12 @@ impl ActionItemCompletionReceipt {
     ) -> Hash {
         let mut hasher = blake3::Hasher::new();
         hasher.update(Self::DOMAIN_TAG);
+        // Use u64 length prefixes to match the canonical encoding used by
+        // `GovernanceProof::compute_proof_hash` and `compute_vote_hash`
+        // elsewhere in this module — keeps the hash binding consistent
+        // across receipt types and avoids any risk of u32 truncation.
         for field in [item_id, domain_id, actor_did] {
-            hasher.update(&(field.len() as u32).to_le_bytes());
+            hasher.update(&(field.len() as u64).to_le_bytes());
             hasher.update(field.as_bytes());
         }
         let transition_byte: u8 = match transition {
